refactor: 도메인 중심 컨트롤러 분리 및 서비스 순환참조 해결
- 기존 UserController, EventController에서 like, visit 관련 엔드포인트를 분리하여
  각각 LikeController, VisitController로 이동
- /users, /events 엔드포인트는 유지하되, 도메인별 컨트롤러에서 담당하도록 구조 개선
- 서비스 간 순환참조 이슈 해결
- 각 컨트롤러는 자신의 도메인 서비스만 의존하도록 변경
- 코드 가독성 및 유지보수성 향상
