+++ src/main/java/com/moonbaar/domain/event/controller/EventController.java
@@ -4,21 +4,12 @@ import com.moonbaar.domain.event.dto.EventDetailResponse;
 import com.moonbaar.domain.event.dto.EventListResponse;
 import com.moonbaar.domain.event.dto.EventSearchRequest;
 import com.moonbaar.domain.event.service.EventService;
-import com.moonbaar.domain.like.dto.LikeResponse;
-import com.moonbaar.domain.like.service.LikeService;
-import com.moonbaar.domain.visit.dto.VisitRequest;
-import com.moonbaar.domain.visit.dto.VisitResponse;
-import com.moonbaar.domain.visit.service.VisitService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -32,8 +23,6 @@ public class EventController {
     private Long MOCK_USER_ID = 1L;
 
     private final EventService eventService;
-    private final LikeService likeService;
-    private final VisitService visitService;
 
     @GetMapping
     public EventListResponse searchEvents(@ModelAttribute @Valid EventSearchRequest request) {
@@ -43,26 +32,5 @@ public class EventController {
     @GetMapping("/{eventId}")
     public EventDetailResponse getEventDetail(@PathVariable Long eventId) {
         return eventService.getEventDetail(eventId);
-    }
-
-    @PostMapping("/{eventId}/like")
-    public ResponseEntity<LikeResponse> likeEvent(@PathVariable Long eventId) {
-        LikeResponse response = likeService.likeEvent(MOCK_USER_ID, eventId);
-        return ResponseEntity.ok(response);
-    }
-
-    @DeleteMapping("/{eventId}/like")
-    public ResponseEntity<LikeResponse> unlikeEvent(@PathVariable Long eventId) {
-        LikeResponse response = likeService.unlikeEvent(MOCK_USER_ID, eventId);
-        return ResponseEntity.ok(response);
-    }
-
-    @PostMapping("/{eventId}/visit")
-    public ResponseEntity<VisitResponse> visitEvent(
-            @PathVariable Long eventId,
-            @Valid @RequestBody VisitRequest request) {
-
-        VisitResponse response = visitService.visitEvent(MOCK_USER_ID, eventId, request);
-        return ResponseEntity.ok(response);
     }
 }
